requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
attrs==23.2.0
Authlib==1.3.1
autopep8==2.3.1
bcrypt==4.1.3
black==24.4.2
bleach==6.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
cssselect==1.2.0
cssutils==2.11.1
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==26.0.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-mail==1.4.1
filelock==3.15.4
flake8==7.1.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pipdeptree==2.23.1
Pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
premailer==3.10.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.2.6
PyMySQL==1.1.1
pytest==8.2.2
pytest-asyncio==0.23.8
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tomli==2.0.1
tomlkit==0.13.0
twilio==9.2.3
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uuid7==0.1.0
uvicorn==0.30.3
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
yarl==1.9.4
stripe==10.7.0
